Test reorganization
This change reorganizes the test into multi files based on the types of functionality being tested. So there's a `multibuild_test.js` and an `export_test.js`. It's easy to find where to put a test by looking in the main `test/test.js` file.

The multibuild_test.js is still quite large and could probably be broken up more, but will defer doing that for now.

Closes #342
