+++ test/order_test.js
@@ -1,0 +1,45 @@
+var comparify = require("comparify"),
+	orderGraph = require("../lib/graph/order");
+
+describe("order", function(){
+
+	it("should work", function(){
+
+		var graph = {
+		   "a":{
+		      "deps":[ "dep_a_b",  "dep_all" ],
+		      "dependencies":[ "dep_a_b", "dep_all" ]
+		   },
+		   "dep_a_b":{
+		      "deps":[],
+		      "dependencies":[]
+		   },
+		   "dep_all":{
+		      "deps":["jquery"],
+		      "dependencies":["jquery"],
+		   },
+		   "jquery": {
+		      deps: [],
+		      dependencies: []
+		   }
+		};
+
+		orderGraph(graph,"a");
+		comparify(graph, {
+		   "a":{
+		      order: 3
+		   },
+		   "dep_a_b":{
+		      order: 0
+		   },
+		   "dep_all":{
+		      order: 2
+		   },
+		   "jquery": {
+		      order: 1
+		   }
+		}, true);
+
+	});
+
+});
